requirements.txt

langchain
langchain-groq
langchain-huggingface
sentence-transformers
gradio==4.44.0
python-dotenv
openai
ipython